requirements.txt
selenium>=4.0.0
PyYAML>=5.4.1
webdriver-manager>=3.5.3